+++ src/std/io/CHANGELOG.md
@@ -1,5 +1,6 @@
 # nstd_io
 ## Latest
+- Added ability to write raw data to stdout and stderr through `nstd_std_io_write_raw[_err]`.
 - Update for Rust 1.57.
 ## 0.2.0
 - `clib` feature.